すべての API を多階層対応
