+++ src/routes/api/[...path].js
@@ -61,7 +61,28 @@ export async function get({params}) {
 /*
  * post
  */
-export async function post({params}) {
+export async function post({params, request}) {
+  let paths = params.path.split('/');
+  let content_id = pathsToContentId(paths);
+  let items = dummy[content_id];
+
+  let generator = dummy.generator[content_id];
+
+  let item = generator();
+  let body = await request.json();
+
+  // id は無視する
+  delete body.item.id;
+  
+  Object.assign(item, body.item);
+
+  items.unshift(item);
+
+  return {
+    body: {
+      item,
+    }
+  };
 };
 
 
@@ -100,5 +121,29 @@ export async function put({params, request}) {
 /*
  * del
  */
-export async function del({params}) {
+export async function del({params, request}) {
+  let paths = params.path.split('/');
+  let content_id = pathsToContentId(paths);
+  let items = dummy[content_id];
+
+  // 単体取得
+  let id = pathsToId(paths);
+  let index = items.findIndex(item => item['id'] === id);
+
+  if (index === -1) {
+    return {
+      status: '404',
+      body: {
+        message: `Not found: ${id}`,
+      }
+    }
+  }
+
+  items.splice(index, 1);
+
+  return {
+    body: {
+      message: `deleted ${params['id']}`,
+    }
+  };
 };
